Cleaned up the NewItemVC Class file and improved the Keyboard functionality
Better Use Swift's of Encapsulation Features
NewItemVC now has 2 extensions: Keyboard Functionality & ImagePicker.
The description TextView in the NewItemVC now has a border around it.
The price field now has a way to end editing, this was done by adding a
“done” button above the keyboard when editing the price field.
Placeholder Images have been updated to better reflect the type of
image they are standing in for.