Requested Meet Up Phrasing updated & Sign out now works
Bug fix:
Sign out previously crashed application. This was dude to having lower
case userSaleCell vs UserSaleCell as the cell identifier in the
UserProfileVC.
Requested meet up changes:
Requested Meet up status now specifies if it was a buyer or a seller.
The Algolia Search Manager now allows the developer to specify which
type of user is sending the meet up request.