More SellerVC updates
SaleItemVC now connected to SellerVC via the “perform segue with
identifier” function. UsernameLabel removed in favor of a username
Button that can be used to trigger the segue.
Minor update to the firebaseManager currently prints the seller’s data
from the snapshot given from Firebase’s database.
Finally Added Constraints to the Seller View’s Elements in the UI
Storyboard.